+++ build.gradle
@@ -29,7 +29,7 @@ test {
 }
 
 application {
-    mainClass.set('seedu.JobPilot.JobPilot')
+    mainClass.set('seedu.jobpilot.jobpilot')
 }
 
 shadowJar {
