+++ src/test/java/task/AddTest.java
@@ -2,8 +2,8 @@ package task;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import seedu.JobPilot.Exceptions.JobPilotException;
-import seedu.JobPilot.JobPilot;
+import exception.JobPilotException;
+import jobpilot.JobPilot;
 import java.util.ArrayList;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
